feat(在线教育前台系统): 完成前台课程模块的课程列表显示(包含根据课程类别/关注度/最新/价格为用户提供不同的课程)以及课程详情(包含课程大纲显示以及阿里云点播视频在线播放功能、课程详情、课程评论功能)，正在开发订单模块用于购买课程(已实现创建订单接口)

+++ guli_edu/service/service_vod/src/main/java/com/geek/vod/controller/VodController.java
@@ -1,6 +1,9 @@
 package com.geek.vod.controller;
 
-import com.geek.commonutils.R;
+import com.aliyuncs.exceptions.ClientException;
+import com.baomidou.mybatisplus.core.toolkit.StringUtils;
+import com.geek.commonutils.result.R;
+import com.geek.servicebase.exception.GuliException;
 import com.geek.vod.service.VodService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiImplicitParam;
@@ -40,6 +43,9 @@ public class VodController {
     @ApiImplicitParam(name = "id",value = "云端视频id")
     @DeleteMapping("removeVideoSourceById/{id}")
     public R removeVideoSourceById(@PathVariable("id") String id){
+        if (StringUtils.isBlank(id)) {
+            throw new GuliException(20001, "视频id为空");
+        }
         return vodService.removeVideoById(id);
     }
 
@@ -49,5 +55,15 @@ public class VodController {
     @DeleteMapping("/deleteBatchVideo")
     public R deleteBatchVideo(@RequestParam("videoSourceIdList") List<String> videoSourceIdList){
         return vodService.deleteBatchVideo(videoSourceIdList);
+    }
+
+    @ApiOperation("根据云端视频id获取视频播放凭证")
+    @GetMapping("getPlayAuth/{videoId}")
+    public R getPlayAuth(@PathVariable("videoId") String videoId) throws ClientException {
+        if (StringUtils.isBlank(videoId)) {
+            throw new GuliException(20001, "视频id为空");
+        }
+        String playAuth = vodService.getPlayAuth(videoId);
+        return R.ok().data("playAuth", playAuth);
     }
 }
